damage system

+++ src/utils/MathUtils.java
@@ -6,7 +6,7 @@ public abstract class MathUtils {
         return (int) Math.floor(Math.random() * (max - min) + min);
     }
 
-    public static int clamp(int min, int max, int val) {
+    public static float clamp(float min, float max, float val) {
         return Math.max(min, Math.min(max, val));
     }
 }
